requirements.txt
setuptools==65.5.1
load_dotenv>=0.1.0
sonnen_api_v2>=0.5.12
